Rename the task and flag to make their purpose clearer
No lock is needed to stop the stream.

Closing the stream is now done by an explicit Microphone.close() which
would only happen if changing the listening source. (Need to check
that keeping it open has no problems if eg, snapclient stops/starts).
(so this means that it's not an error anymore)

Signed-off-by: David Greaves <david@dgreaves.com>